requirements.txt
flask
requests
beautifulsoup4
openai
python-dotenv
pandas